docs(react-router): Fix react router v3 example
+ move to react-router-v3 folder

fixes #171
