+++ packages/react-instantsearch/examples/react-router-v3/README.md
