+++ packages/react-instantsearch/examples/react-router-v3/package.json
@@ -4,14 +4,14 @@
     "react-scripts": "^1.0.10"
   },
   "dependencies": {
-    "lodash": "^4.16.6",
-    "prop-types": "^15.5.8",
-    "qs": "^6.3.0",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
+    "lodash": "^4.17.4",
+    "prop-types": "^15.5.10",
+    "qs": "^6.5.0",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
     "react-instantsearch": "beta",
     "react-instantsearch-theme-algolia": "beta",
-    "react-router-dom": "^4.1.2"
+    "react-router": "^3.0.5"
   },
   "scripts": {
     "start": "react-scripts start"
