+++ packages/react-instantsearch/examples/react-router-v3/public/index.html
