+++ packages/react-instantsearch/examples/react-router-v3/src/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { Router, Route, browserHistory } from 'react-router';
 
 ReactDOM.render(
-  <Router>
+  <Router history={browserHistory}>
     <Route path="/" component={App} />
   </Router>,
   document.getElementById('root')
