+++ packages/react-instantsearch/examples/react-router/src/App.js
@@ -1,4 +1,3 @@
-import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import {
   InstantSearch,
@@ -12,41 +11,36 @@ import {
   SearchBox,
   ClearAll,
 } from 'react-instantsearch/dom';
-import { withRouter } from 'react-router';
 import 'react-instantsearch-theme-algolia/style.css';
+import PropTypes from 'prop-types';
 import qs from 'qs';
-import { isEqual } from 'lodash';
+
+const updateAfter = 700;
+
+const createURL = state => `?${qs.stringify(state)}`;
+
+const searchStateToUrl = (props, searchState) =>
+  searchState ? `${props.location.pathname}${createURL(searchState)}` : '';
 
 class App extends Component {
   constructor(props) {
     super(props);
-    this.state = { searchState: { ...qs.parse(props.router.location.query) } };
+    this.state = { searchState: qs.parse(props.location.search.slice(1)) };
   }
 
-  componentWillReceiveProps() {
-    this.setState({ searchState: qs.parse(this.props.router.location.query) });
-  }
-
-  shouldComponentUpdate(nextProps, nextState) {
-    return !isEqual(this.state.searchState, nextState.searchState);
-  }
-
-  onSearchStateChange(nextSearchState) {
-    const THRESHOLD = 700;
-    const newPush = Date.now();
-    this.setState({ lastPush: newPush, searchState: nextSearchState });
-    if (this.state.lastPush && newPush - this.state.lastPush <= THRESHOLD) {
-      this.props.router.replace(
-        nextSearchState ? `?${qs.stringify(nextSearchState)}` : ''
-      );
-    } else {
-      this.props.router.push(
-        nextSearchState ? `?${qs.stringify(nextSearchState)}` : ''
-      );
-    }
-  }
-
-  createURL = state => `?${qs.stringify(state)}`;
+  onSearchStateChange = searchState => {
+    clearTimeout(this.debouncedSetState);
+    this.debouncedSetState = setTimeout(
+      () => {
+        this.props.history.push(
+          searchStateToUrl(this.props, searchState),
+          searchState
+        );
+      },
+      updateAfter
+    );
+    this.setState({ searchState });
+  };
 
   render() {
     return (
@@ -56,7 +50,7 @@ class App extends Component {
         indexName="ikea"
         searchState={this.state.searchState}
         onSearchStateChange={this.onSearchStateChange.bind(this)}
-        createURL={this.createURL.bind(this)}
+        createURL={createURL}
       >
 
         <div>
@@ -106,7 +100,10 @@ class App extends Component {
 }
 
 App.propTypes = {
-  router: PropTypes.object.isRequired,
+  history: PropTypes.shape({
+    push: PropTypes.func.isRequired,
+  }),
+  location: PropTypes.object.isRequired,
 };
 
-export default withRouter(App);
+export default App;
